build-tools: Rename src/lib to src/library
