+++ build-tools/packages/build-cli/src/commands/check/changeset.ts
@@ -7,7 +7,7 @@ import chalk from "chalk";
 import { sortPackageJson as sortJson } from "sort-package-json";
 
 import { BaseCommand } from "../../base";
-import { Repository } from "../../lib";
+import { Repository } from "../../library";
 
 export default class CheckChangesetCommand extends BaseCommand<typeof CheckChangesetCommand> {
 	static readonly summary = `Checks if a changeset was added when compared against a branch. This is used in CI to enforce that changesets are present for a PR.`;
