+++ build-tools/packages/build-cli/src/commands/release/fromTag.ts
@@ -8,7 +8,7 @@ import { Args } from "@oclif/core";
 import semver from "semver";
 import { sortPackageJson as sortJson } from "sort-package-json";
 
-import { sortVersions } from "../../lib";
+import { sortVersions } from "../../library";
 import { ReleaseGroup, ReleasePackage } from "../../releaseGroups";
 import { ReleaseReportBaseCommand, ReleaseSelectionMode } from "./report";
 import { findPackageOrReleaseGroup } from "../../args";
