+++ build-tools/packages/build-cli/src/handlers/doFunctions.ts
@@ -10,7 +10,7 @@ import { FluidRepo, MonoRepo } from "@fluidframework/build-tools";
 
 import { bumpVersionScheme, detectVersionScheme } from "@fluid-tools/version-tools";
 
-import { difference, getPreReleaseDependencies, npmCheckUpdates, setVersion } from "../lib";
+import { difference, getPreReleaseDependencies, npmCheckUpdates, setVersion } from "../library";
 import { CommandLogger } from "../logging";
 import { MachineState } from "../machines";
 import { ReleaseGroup, ReleasePackage, isReleaseGroup } from "../releaseGroups";
