+++ build-tools/packages/build-cli/test/library/changesets.test.ts
@@ -6,7 +6,7 @@ import { assert, expect } from "chai";
 import { pathExistsSync } from "fs-extra";
 import path from "node:path";
 
-import { flattenChangesets, groupByPackage, loadChangesets } from "../../src/lib/changesets";
+import { flattenChangesets, groupByPackage, loadChangesets } from "../../src/library/changesets";
 
 const changesetsPath = path.resolve(__dirname, "../data");
 assert.isTrue(pathExistsSync(changesetsPath));
